StoreStatusAction: support no_overwrite to skip fetch
When no_overwrite is true and the S3 object already exists, skip
fxtwitter fetch and S3 put. Extract quoted_status_id from existing
data (symbol keys) instead. Include no_overwrite in returned hash
so SFN can carry it forward.

+++ collector/lib/publikes/store_status_action.rb
@@ -4,21 +4,23 @@ require 'aws-sdk-s3'
 
 module Publikes
   class StoreStatusAction
-    def initialize(environment:, status_id:, visited_status_ids: [])
+    def initialize(environment:, status_id:, visited_status_ids: [], no_overwrite: false)
       @environment = environment
       @status_id = status_id.to_s
       @visited_status_ids = visited_status_ids.map(&:to_s)
+      @no_overwrite = no_overwrite
 
       raise ArgumentError, "invalid status_id" unless @status_id.match?(/\A[0-9a-zA-Z]+\z/)
     end
 
-    attr_reader :status_id, :visited_status_ids
+    attr_reader :status_id, :visited_status_ids, :no_overwrite
     def env; @environment; end
 
     USER_AGENT = 'Publikes-Crawler (+https://github.com/sorah/publikes)'
 
     def perform
       key = "data/private/statuses/#{status_id}.json"
+      existing = true
       data = begin
         JSON.parse(
           env.s3.get_object(
@@ -28,6 +30,7 @@ module Publikes
           symbolize_names: true,
         )
       rescue Aws::S3::Errors::NoSuchKey
+        existing = false
         {
           id: status_id.to_s,
           complete: false,
@@ -37,30 +40,35 @@ module Publikes
         }
       end
 
-      fxtwitter_data = begin
-        JSON.parse(URI.open("https://api.fxtwitter.com/status/#{status_id}", "User-Agent" => USER_AGENT, &:read))
-      rescue OpenURI::HTTPError => e
-        code = e.io.status[0]
-        raise unless code == '404' || code == '403' || code == '401'
+      if no_overwrite && existing
+        quoted_status_id = data[:fxtwitter_data]&.dig(:tweet, :quote, :id)&.to_s
+      else
+        fxtwitter_data = begin
+          JSON.parse(URI.open("https://api.fxtwitter.com/status/#{status_id}", "User-Agent" => USER_AGENT, &:read))
+        rescue OpenURI::HTTPError => e
+          code = e.io.status[0]
+          raise unless code == '404' || code == '403' || code == '401'
+        end
+
+        new_data = data.merge(
+          complete: true,
+          saved_at: Time.now.to_i,
+          fxtwitter_data: fxtwitter_data || data[:fxtwitter_data],
+        )
+
+        env.s3.put_object(
+          bucket: env.s3_bucket,
+          key:,
+          content_type: "application/json; charset=utf-8",
+          body: JSON.generate(new_data),
+        )
+
+        quoted_status_id = fxtwitter_data&.dig('tweet', 'quote', 'id')&.to_s
       end
 
-      new_data = data.merge(
-        complete: true,
-        saved_at: Time.now.to_i,
-        fxtwitter_data: fxtwitter_data || data[:fxtwitter_data],
-      )
-
-      env.s3.put_object(
-        bucket: env.s3_bucket,
-        key:,
-        content_type: "application/json; charset=utf-8",
-        body: JSON.generate(new_data),
-      )
-
-      quoted_status_id = fxtwitter_data&.dig('tweet', 'quote', 'id')&.to_s
       new_visited = visited_status_ids | [status_id]
 
-      result = { status_id:, visited_status_ids: new_visited }
+      result = { status_id:, visited_status_ids: new_visited, no_overwrite: }
       if quoted_status_id && !new_visited.include?(quoted_status_id)
         result[:quoted_status_id] = quoted_status_id
       end
